==> 1_create_input_dataset.ipynb <==
import pandas as pd
df = pd.read_csv('./data/ensembled_annotations_public.csv')

from datasets import load_dataset

DATASET = "princeton-nlp/SWE-bench"
ds = load_dataset(DATASET, split="test")
_d = pd.DataFrame(ds)

input_bouncer_data = pd.merge(df, _d, on="instance_id", how="inner")
input_bouncer_data["input_bounce"] = input_bouncer_data["underspecified"] >= 2
input_bouncer_data["input_quality"] = input_bouncer_data["underspecified"]

input_bouncer_data.to_csv("./dataset/input_bouncer.csv", index=False)

input_bouncer_data["underspecified"].value_counts()

input_bouncer_data["input_bounce"].value_counts()
